new BankTransaction xml format

+++ trunk/src/main/java/ua/org/tumakha/spdtool/services/BankTransactionService.java
@@ -17,7 +17,7 @@ public interface BankTransactionService {
 
 	BankTransaction findTransaction(Integer transactionId);
 
-	BankTransaction findTransactionByUserAndId(Integer userId, Integer id);
+	BankTransaction findTransactionByUserAndId(Integer userId, Long id);
 
 	List<BankTransaction> findAllTransactions();
